# docs/openapi/fanta-win-api-internal.yaml
openapi: 3.0.3
info:
  title: API B2B Gestione Rosa e Predizione Formazione
  description: >-
    API REST B2B per la gestione della rosa su DynamoDB e il calcolo predittivo 
    della formazione ottimale di Fantacalcio (FantaClaude).
  version: 2.12.0
  x-api-id: fantaclaude-lineup-manager-api-b2b
  termsOfService: https://fantaclaude.ai/termini-di-servizio
  contact:
    name: Supporto FantaClaude
    url: https://fantaclaude.ai/supporto
servers:
  - url: http://localhost:8081
    description: Ambiente locale di sviluppo (server.port=8081, nessun context-path configurato)
  - url: https://api.fantaclaude.ai
    description: Ambiente di produzione (placeholder, non ancora deployato)
...
paths:
...
  /fanta-private/matchstats/import:
    post:
      summary: Importa le statistiche di una giornata da file CSV
      description: >
        Carica un file CSV con voto/fantavoto/bonus/malus dei calciatori per una giornata gia' disputata
        (fonte: sito voti fantacalcio, compilazione manuale). Formato atteso, separatore ';', UTF-8, intestazione
        su prima riga: giocatore;squadra;giornata;stagione;avversario;casa;voto;fantavoto;gol;assist;ammonizioni;espulsioni;xg;xa.
        Il giocatore viene risolto per nome+squadra reale contro l'anagrafica; le righe non risolvibili
        non bloccano l'import ma vengono riportate in errors.
      operationId: importMatchStats
      tags:
        - MatchStats
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
                  description: File CSV con le statistiche della giornata
      responses:
        '200':
          description: Import completato (anche parzialmente, vedi errors)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MatchStatsImportResult'
        '400':
          description: File assente o non leggibile come CSV
          content:
            application/json:
              schema:
                $ref: 'problem-internal.yaml#/components/schemas/Problem'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: 'problem-internal.yaml#/components/schemas/Problem'
        '503':
          description: Storage DynamoDB temporaneamente non raggiungibile
          content:
            application/json:
              schema:
                $ref: 'problem-internal.yaml#/components/schemas/Problem'
  /fanta-private/matchstats/fixture-lookup/{realTeam}/{matchday}:
    get:
      summary: Recupera da Football-Data.org il vero avversario di una squadra per una giornata
      description: >
        Interroga in tempo reale Football-Data.org (lo stesso client esterno usato dalla sincronizzazione
        delle Fixtures) per restituire avversario reale, casa/trasferta, data e stato della partita di
        una squadra per una specifica giornata di Serie A. Pensato per verificare/prefillare a mano i
        campi avversario/casa prima di compilare il CSV di /fanta-private/matchstats/import, evitando
        errori di trascrizione (es. avversario scritto uguale alla squadra stessa).
      operationId: lookupMatchFixture
      tags:
        - MatchStats
      parameters:
        - name: realTeam
          in: path
          required: true
          description: Nome della squadra reale, come in anagrafica giocatori
          schema:
            type: string
            example: Inter
        - name: matchday
          in: path
          required: true
          description: Numero della giornata di Serie A da cercare
          schema:
            type: integer
            example: 4
      responses:
        '200':
          description: Partita trovata
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FixtureLookupResult'
        '404':
          description: >-
            Nessuna partita trovata per questa squadra/giornata, oppure la squadra non e' risolvibile
            su Football-Data.org
          content:
            application/json:
              schema:
                $ref: 'problem-internal.yaml#/components/schemas/Problem'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: 'problem-internal.yaml#/components/schemas/Problem'
        '503':
          description: Football-Data.org temporaneamente non raggiungibile
          content:
            application/json:
              schema:
                $ref: 'problem-internal.yaml#/components/schemas/Problem'
...
  /fanta-private/matchstats/sync-real/{rosterId}/{matchday}:
    post:
      summary: Sincronizza da SportMonks gol/assist/cartellini/gol-subiti reali di una giornata
      description: >
        Per ogni giocatore in rosa, legge da SportMonks (endpoint /rounds/seasons/{seasonId}, un'unica
        chiamata per l'intera giornata) gli eventi oggettivi della partita della sua squadra reale in
        quella giornata: gol, assist (related_player_id sull'evento Goal), ammonizioni, espulsioni e,
        per portieri/difensori, gol subiti dalla squadra. Il voto (pagella) NON viene toccato: resta
        quello gia' presente (manuale, es. da /fanta-private/matchstats/import) o vuoto se il giocatore
        non ha ancora un voto per questa giornata - in quel caso fantavoto resta vuoto finche' il voto
        non viene valorizzato. Se la riga PlayerMatchStat per playerId+giornata non esiste ancora viene
        creata. Richiede fantawin.sportmonks-serie-a-season-id configurato. Al termine scrive anche un
        export CSV con lo storico corrente del roster in fantawin.match-stats-export-dir.
      operationId: syncRealMatchStats
      tags:
        - MatchStats
      parameters:
        - name: rosterId
          in: path
          required: true
          description: id della squadra del fantacalcio
          schema:
            type: string
            example: junior
        - name: matchday
          in: path
          required: true
          description: numero di giornata di Serie A da sincronizzare
          schema:
            type: integer
            example: 5
      responses:
        '200':
          description: Sincronizzazione eseguita (anche se qualche partita non e' stata trovata)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MatchStatsSyncResult'
        '404':
          description: Rosa non trovata
          content:
            application/json:
              schema:
                $ref: 'problem-internal.yaml#/components/schemas/Problem'
        '500':
          description: Internal server error (es. fantawin.sportmonks-serie-a-season-id non configurato)
          content:
            application/json:
              schema:
                $ref: 'problem-internal.yaml#/components/schemas/Problem'
        '503':
          description: SportMonks temporaneamente non raggiungibile
          content:
            application/json:
              schema:
                $ref: 'problem-internal.yaml#/components/schemas/Problem'
  /fanta-private/predict/lineup/{rosterId}:
    post:
      summary: Calcola l'undici titolare e la panchina ottimale per la giornata
      description: >
        Elabora i dati attuali della rosa (stato di forma, storico avversario, impegni europei/infrasettimanali e titolarità)
        e applica i moduli logici del FantaRating per calcolare la formazione ideale e il Modificatore di Difesa.
      operationId: calculateOptimalLineup
      tags:
        - Prediction
      parameters:
        - name: rosterId
          in: path
          required: true
          description: id della squadra del fantacalcio
          schema:
            type: string
            example: junior
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LineupRequest'
      responses:
        '200':
          description: Calcolo predittivo completato con successo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LineupResponse'
        '422':
          description: >-
            Rosa insufficiente per comporre uno qualsiasi dei moduli previsti (3-4-3, 4-3-3, 3-5-2,
            4-4-2) - es. nessun portiere disponibile o troppi pochi giocatori schierabili
          content:
            application/json:
              schema:
                $ref: 'problem-internal.yaml#/components/schemas/Problem'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: 'problem-internal.yaml#/components/schemas/Problem'
        '503':
          description: Storage DynamoDB temporaneamente non raggiungibile
          content:
            application/json:
              schema:
                $ref: 'problem-internal.yaml#/components/schemas/Problem'
...
components:
  schemas:
...
    MatchStatsImportResult:
      type: object
      required:
        - imported
        - skipped
        - errors
      properties:
        imported:
          type: integer
          description: Numero di righe importate/aggiornate con successo
          example: 22
        skipped:
          type: integer
          description: Numero di righe scartate (giocatore non risolto o riga non valida)
          example: 1
        errors:
          type: array
          description: Dettaglio delle righe scartate, con numero di riga e motivo
          items:
            type: string
          example: ["riga 5: nessun giocatore trovato per 'Mario Rossi' (Empoli)"]
...
    MatchStatsSyncResult:
      type: object
      required:
        - updatedPlayers
        - unresolvedTeams
      properties:
        updatedPlayers:
          type: array
          items:
            $ref: '#/components/schemas/PlayerMatchStatUpdate'
        unresolvedTeams:
          type: array
          description: >-
            Squadre reali per cui non e' stato possibile trovare la fixture di questa giornata su
            SportMonks (round non trovato, fixture ambigua/assente o nome squadra non risolto tra i
            participants)
          items:
            type: string
          example: [Milan]
    PlayerMatchStatUpdate:
      type: object
      required:
        - playerId
        - name
        - realTeam
        - gol
        - assist
        - ammonizioni
        - espulsioni
      properties:
        playerId:
          type: string
        name:
          type: string
          example: Barella
        realTeam:
          type: string
          example: Inter
        gol:
          type: integer
        assist:
          type: integer
        ammonizioni:
          type: integer
        espulsioni:
          type: integer
        golSubiti:
          type: integer
          nullable: true
          description: Valorizzato solo per portieri/difensori
        fantavoto:
          type: number
          format: double
          nullable: true
          description: Null se il voto non e' ancora stato inserito manualmente per questa giornata
    FixtureLookupResult:
      type: object
      required:
        - realTeam
        - opponentTeam
        - home
        - matchday
        - status
      properties:
        realTeam:
          type: string
          description: Squadra reale richiesta
          example: Inter
        opponentTeam:
          type: string
          description: Vero avversario di realTeam per la giornata richiesta, da Football-Data.org
          example: Udinese
        home:
          type: boolean
          description: true se realTeam gioca in casa
          example: true
        matchday:
          type: integer
          example: 4
        utcDate:
          type: string
          description: Data/ora UTC della partita (ISO-8601), cosi' come restituita da Football-Data.org
          example: "2026-09-14T18:30:00Z"
        status:
          type: string
          description: Stato della partita su Football-Data.org (es. FINISHED, SCHEDULED, TIMED)
          example: FINISHED
...
    Player:
      type: object
      required:
        - id
        - name
        - realTeam
        - position
        - status
        - nextOpponent
        - startingProbability
        - recentScores
      properties:
        id:
          type: string
          description: id del giocatore
          example: PLAYER#inter_lautaro
        name:
          type: string
          example: Lautaro Martinez
        realTeam:
          type: string
          example: Inter
        position:
          type: string
          enum: [POR, DIF, CEN, ATT]
          example: ATT
        status:
          type: string
          enum: [DISPONIBILE, IN_DUBBIO, INFORTUNATO, SQUALIFICATO]
          example: DISPONIBILE
        nextOpponent:
          type: string
          example: Empoli
        startingProbability:
          type: number
          format: float
          description: Probabilità di titolarità in percentuale (0-100)
          example: 85.0
        recentScores:
          type: array
          description: Ultimi voti/fantamedia ottenuti nelle ultime 3-5 partite
          items:
            type: number
            format: float
          example: [7.5, 6.0, 10.0]
    LineupRequest:
      type: object
      required:
        - matchday
        - modifierActive
      properties:
        matchday:
          type: integer
          description: Numero della giornata di Serie A
          example: 8
        modifierActive:
          type: boolean
          description: Abilita il calcolo del Modificatore di Difesa
          example: true
        teamsWithMidweekMatches:
          $ref: 'lineup-shared-internal.yaml#/components/schemas/TeamsWithMidweekMatchesOverride'
    LineupResponse:
      type: object
      required:
        - winningFormation
        - totalExpectedScore
        - defenseModifierBonus
        - startingEleven
        - bench
      properties:
        winningFormation:
          type: string
          description: Modulo tattico ottimale selezionato (3-4-3, 4-3-3, 3-5-2, 4-4-2)
          example: 4-3-3
        totalExpectedScore:
          type: number
          format: float
          description: Punteggio totale atteso calcolato dall'algoritmo
          example: 74.8
        defenseModifierBonus:
          type: integer
          description: Bonus aggiunto dal Modificatore di Difesa (+0, +1, +3, +6)
          example: 3
        startingEleven:
          type: array
          items:
            $ref: '#/components/schemas/PlayerRatingDetails'
        bench:
          type: array
          items:
            $ref: '#/components/schemas/PlayerRatingDetails'
    PlayerRatingDetails:
      type: object
      required:
        - player
        - fantaRating
        - fantaRatingBreakdown
      properties:
        player:
          $ref: '#/components/schemas/Player'
        fantaRating:
          type: number
          format: float
          description: Rating finale di schierabilità (da 1 a 10)
          example: 8.4
        fantaRatingBreakdown:
          $ref: '#/components/schemas/FantaRatingBreakdown'
    FantaRatingBreakdown:
      type: object
      required:
        - baseFormScore
        - historicalScore
        - matchDifficultyXgScore
        - midweekPenalty
        - availabilityPenalty
      properties:
        baseFormScore:
          type: number
          format: float
          description: Punteggio dal 40% sul recente stato di forma
          example: 3.2
        historicalScore:
          type: number
          format: float
          description: Punteggio dal 30% sullo storico contro l'avversario
          example: 2.1
        matchDifficultyXgScore:
          type: number
          format: float
          description: Punteggio dal 30% sulla difficoltà del match e xG/xA
          example: 2.5
        midweekPenalty:
          type: number
          format: float
          description: Penalità applicata per partite infrasettimanali/coppe
          example: -0.8
        availabilityPenalty:
          type: number
          format: float
          description: Penalità applicata per ballottaggi o dubbio titolarità
          example: 0.0